# backendserver/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ticketing API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://0b44d62428f0.ngrok-free.app",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/events": {
      "get": {
        "summary": "List Events",
        "operationId": "list_events_events_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventRead"
                  },
                  "type": "array",
                  "title": "Response List Events Events Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Event",
        "operationId": "create_event_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/playlist": {
      "post": {
        "summary": "Generate Playlist",
        "operationId": "generate_playlist_events__event_id__playlist_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/payments/checkout": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Create Checkout Session",
        "operationId": "create_checkout_session_payments_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payments/confirm/{session_id}": {
      "get": {
        "tags": [
          "payments"
        ],
        "summary": "Confirm Payment",
        "operationId": "confirm_payment_payments_confirm__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/integrations/google/calendar": {
      "post": {
        "summary": "Add Calendar Event",
        "description": "Creates a Google Calendar event using the provided token.",
        "operationId": "add_calendar_event_integrations_google_calendar_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Optional refresh token",
              "title": "Refresh Token"
            },
            "description": "Optional refresh token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/google/gmail": {
      "post": {
        "summary": "Send Gmail",
        "description": "Sends an email using the system's SMTP configuration.\nNo Google Auth required from the user.",
        "operationId": "send_gmail_integrations_google_gmail_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/integrations/google/gmail/by-booking/{booking_id}": {
      "post": {
        "summary": "Send Gmail By Booking",
        "description": "Sends an email using the system's SMTP configuration.\nDoes NOT require the user to have completed Google OAuth.",
        "operationId": "send_gmail_by_booking_integrations_google_gmail_by_booking__booking_id__post",
        "parameters": [
          {
            "name": "booking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Booking Id"
            }
          },
          {
            "name": "to_email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Recipient email address",
              "title": "To Email"
            },
            "description": "Recipient email address"
          },
          {
            "name": "subject",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Email subject",
              "title": "Subject"
            },
            "description": "Email subject"
          },
          {
            "name": "body",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Email body",
              "title": "Body"
            },
            "description": "Email body"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BookingCreate": {
        "properties": {
          "event_id": {
            "type": "integer",
            "title": "Event Id"
          },
          "customer_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Customer Name"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "customer_name",
          "quantity"
        ],
        "title": "BookingCreate"
      },
...
      "CalendarEventRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "title": "Time"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "name",
          "location",
          "date",
          "time",
          "token"
        ],
        "title": "CalendarEventRequest"
      },
      "EmailRequest": {
        "properties": {
          "to_email": {
            "type": "string",
            "title": "To Email"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "to_email",
          "subject",
          "body",
          "token"
        ],
        "title": "EmailRequest"
      },
      "EventCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "format": "time",
            "title": "Time"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Location"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Capacity"
          },
          "tickets_sold": {
            "type": "integer",
            "title": "Tickets Sold",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "date",
          "time",
          "location",
          "price",
          "capacity"
        ],
        "title": "EventCreate"
      },
      "EventRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "format": "time",
            "title": "Time"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "location": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Location"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "capacity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Capacity"
          },
          "tickets_sold": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Tickets Sold",
            "default": 0
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "map_link": {
            "type": "string",
            "title": "Map Link",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "date",
          "time",
          "location",
          "price",
          "capacity",
          "id",
          "map_link"
        ],
        "title": "EventRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}